auth/dev-oidc/requirements.txt
fastapi
uvicorn
itsdangerous
python-jose[cryptography]
python-multipart
